fix(ci-cd-code-reviewer): drop .int() from model-boundary score field
- Zod v4's z.number().int() emits {minimum: -MAX_SAFE_INT, maximum: MAX_SAFE_INT} in its JSON Schema output, which Azure-hosted OpenRouter models reject
- Model boundary now uses bare z.number(); the strict reviewSchema still enforces integer + 1..10 range post-generation
- Verified emitted JSON Schema has zero suspect keywords (minimum, maximum, minItems, maxItems, minLength, maxLength, pattern)
- 61 tests pass

+++ packages/code-reviewer/src/schema.ts
@@ -38,7 +38,10 @@ export type Review = z.infer<typeof reviewSchema>;
 // generation, wrapped as LLM_INVALID_OUTPUT on failure.
 const PermissiveCriterionSchema = z.object({
   name: CriterionName,
-  score: z.number().int(),
+  // z.number() alone emits {type:"number"} with no bounds. z.number().int()
+  // adds minimum/maximum safe-integer bounds which Azure rejects. Strict
+  // int/range enforcement lives in reviewSchema (validated post-generation).
+  score: z.number(),
   rationale: z.string(),
 });
